Update README.md
Updated the GDS section with Degree Centrality, .write() methods for all the algos, and a query to show a small component that was identified.  Fixed a dangling -[:NAMES]-> relationship in one of the cypher queries.  Removed the call to apoc.algos.dijkstra (want to do this in GDS).  Changed the route for shortest path to Bruxelles Midi to London St. Pancras